+++ packages/google_mobile_ads/ios/Tests/FLTRewardedInterstitialAdTest.m
@@ -34,16 +34,8 @@
   OCMStub([request keywords]).andReturn(@[ @"apple" ]);
   GADRequest *gadRequest = OCMClassMock([GADRequest class]);
   OCMStub([request asGADRequest:[OCMArg any]]).andReturn(gadRequest);
-  FLTServerSideVerificationOptions *serverSideVerificationOptions =
-      OCMClassMock([FLTServerSideVerificationOptions class]);
-  GADServerSideVerificationOptions *gadOptions =
-      OCMClassMock([GADServerSideVerificationOptions class]);
-  OCMStub([serverSideVerificationOptions asGADServerSideVerificationOptions])
-      .andReturn(gadOptions);
 
-  [self testLoadShowRewardedInterstitialAd:request
-                           gadOrGAMRequest:gadRequest
-             serverSideVerificationOptions:serverSideVerificationOptions];
+  [self testLoadShowRewardedInterstitialAd:request gadOrGAMRequest:gadRequest];
 }
 
 - (void)testLoadShowRewardedInterstitialAdGAMRequest {
@@ -51,30 +43,19 @@
   OCMStub([request keywords]).andReturn(@[ @"apple" ]);
   GAMRequest *gamRequest = OCMClassMock([GAMRequest class]);
   OCMStub([request asGAMRequest:[OCMArg any]]).andReturn(gamRequest);
-  FLTServerSideVerificationOptions *serverSideVerificationOptions =
-      OCMClassMock([FLTServerSideVerificationOptions class]);
-  GADServerSideVerificationOptions *gadOptions =
-      OCMClassMock([GADServerSideVerificationOptions class]);
-  OCMStub([serverSideVerificationOptions asGADServerSideVerificationOptions])
-      .andReturn(gadOptions);
 
-  [self testLoadShowRewardedInterstitialAd:request
-                           gadOrGAMRequest:gamRequest
-             serverSideVerificationOptions:serverSideVerificationOptions];
+  [self testLoadShowRewardedInterstitialAd:request gadOrGAMRequest:gamRequest];
 }
 
 // Helper method for testing with FLTAdRequest and FLTGAMAdRequest.
 - (void)testLoadShowRewardedInterstitialAd:(FLTAdRequest *)request
-                           gadOrGAMRequest:(GADRequest *)gadOrGAMRequest
-             serverSideVerificationOptions:
-                 (FLTServerSideVerificationOptions *)options {
+                           gadOrGAMRequest:(GADRequest *)gadOrGAMRequest {
   UIViewController *mockRootViewController =
       OCMClassMock([UIViewController class]);
   FLTRewardedInterstitialAd *ad =
       [[FLTRewardedInterstitialAd alloc] initWithAdUnitId:@"testId"
                                                   request:request
                                        rootViewController:mockRootViewController
-                            serverSideVerificationOptions:options
                                                      adId:@1];
   ad.manager = mockManager;
 
@@ -122,11 +103,6 @@
         [invocation getArgument:&rewardHandler atIndex:3];
         rewardHandler();
       });
-  // Prevent calls to setServerSideVerificationOptions if options is nil.
-  if (options == nil) {
-    OCMReject([rewardedInterstitialClassMock
-        setServerSideVerificationOptions:[OCMArg any]]);
-  }
 
   // Mock callback of paid event handler.
   GADAdValue *adValue = OCMClassMock([GADAdValue class]);
@@ -150,12 +126,6 @@
                        responseInfo:[OCMArg isEqual:responseInfo]]);
   OCMVerify([rewardedInterstitialClassMock
       setFullScreenContentDelegate:[OCMArg isEqual:ad]]);
-  if (options != nil) {
-    GADServerSideVerificationOptions *gadOptions =
-        [options asGADServerSideVerificationOptions];
-    OCMVerify([rewardedInterstitialClassMock
-        setServerSideVerificationOptions:[OCMArg isEqual:gadOptions]]);
-  }
   OCMVerify([mockManager
       onPaidEvent:[OCMArg isEqual:ad]
             value:[OCMArg checkWithBlock:^BOOL(id obj) {
@@ -168,6 +138,20 @@
               return TRUE;
             }]]);
 
+  // Set SSV and verify interactions with mocks
+  FLTServerSideVerificationOptions *serverSideVerificationOptions =
+      OCMClassMock([FLTServerSideVerificationOptions class]);
+  GADServerSideVerificationOptions *gadOptions =
+      OCMClassMock([GADServerSideVerificationOptions class]);
+  OCMStub([serverSideVerificationOptions asGADServerSideVerificationOptions])
+      .andReturn(gadOptions);
+
+  [ad setServerSideVerificationOptions:serverSideVerificationOptions];
+
+  OCMVerify([rewardedInterstitialClassMock
+      setServerSideVerificationOptions:[OCMArg isEqual:gadOptions]]);
+
+  // Show the ad and verify callbacks are invoked
   [ad show];
 
   OCMVerify([rewardedInterstitialClassMock
@@ -221,7 +205,6 @@
       [[FLTRewardedInterstitialAd alloc] initWithAdUnitId:@"testId"
                                                   request:request
                                        rootViewController:mockRootViewController
-                            serverSideVerificationOptions:nil
                                                      adId:@1];
   ad.manager = mockManager;
 
@@ -247,16 +230,6 @@
       completionHandler:[OCMArg any]]));
   OCMVerify([mockManager onAdFailedToLoad:[OCMArg isEqual:ad]
                                     error:[OCMArg isEqual:error]]);
-}
-
-- (void)testServerSideVerificationOptionsNil {
-  FLTAdRequest *request = OCMClassMock([FLTAdRequest class]);
-  OCMStub([request keywords]).andReturn(@[ @"apple" ]);
-  GADRequest *gadRequest = OCMClassMock([GADRequest class]);
-  OCMStub([request asGADRequest:[OCMArg any]]).andReturn(gadRequest);
-  [self testLoadShowRewardedInterstitialAd:request
-                           gadOrGAMRequest:gadRequest
-             serverSideVerificationOptions:nil];
 }
 
 @end