+++ packages/google_mobile_ads/lib/src/ad_containers.dart
@@ -13,7 +13,6 @@
 // limitations under the License.
 
 import 'dart:async';
-import 'dart:io' show Platform;
 import 'dart:math';
 
 import 'package:flutter/foundation.dart';
@@ -21,6 +20,7 @@ import 'package:flutter/gestures.dart';
 import 'package:flutter/rendering.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter/widgets.dart';
+import 'package:visibility_detector/visibility_detector.dart';
 
 import 'ad_instance_manager.dart';
 import 'ad_listeners.dart';
@@ -143,40 +143,6 @@ class LoadAdError extends AdError {
   }
 }
 
-/// Location parameters that can be configured in an ad request.
-class LocationParams {
-  /// Location parameters that can be configured in an ad request.
-  const LocationParams({
-    required this.accuracy,
-    required this.longitude,
-    required this.latitude,
-    this.time,
-  });
-
-  /// The accuracy in meters.
-  final double accuracy;
-
-  /// The longitude in degrees.
-  final double longitude;
-
-  /// The latitude in degrees.
-  final double latitude;
-
-  /// The UTC time, in milliseconds since epoch (January 1, 1970).
-  ///
-  /// This is required on Android, and ignored on iOS.
-  final int? time;
-
-  @override
-  bool operator ==(Object other) {
-    return other is LocationParams &&
-        accuracy == other.accuracy &&
-        longitude == other.longitude &&
-        latitude == other.latitude &&
-        time == other.time;
-  }
-}
-
 /// Targeting info per the AdMob API.
 ///
 /// This class's properties mirror the native AdRequest API. See for example:
@@ -189,7 +155,6 @@ class AdRequest {
     this.neighboringContentUrls,
     this.nonPersonalizedAds,
     this.httpTimeoutMillis,
-    this.location,
     this.mediationExtrasIdentifier,
     this.extras,
   });
@@ -215,12 +180,6 @@ class AdRequest {
   ///
   /// This is only supported in Android. This value is ignored on iOS.
   final int? httpTimeoutMillis;
-
-  /// Location data.
-  ///
-  /// Used for mediation targeting purposes.
-  @Deprecated('Location is not used and will be deleted in a future release')
-  final LocationParams? location;
 
   /// String identifier used in providing mediation extras.
   ///
@@ -258,7 +217,6 @@ class AdManagerAdRequest extends AdRequest {
     bool? nonPersonalizedAds,
     int? httpTimeoutMillis,
     this.publisherProvidedId,
-    LocationParams? location,
     String? mediationExtrasIdentifier,
     Map<String, String>? extras,
   }) : super(
@@ -267,7 +225,6 @@ class AdManagerAdRequest extends AdRequest {
           neighboringContentUrls: neighboringContentUrls,
           nonPersonalizedAds: nonPersonalizedAds,
           httpTimeoutMillis: httpTimeoutMillis,
-          location: location,
           mediationExtrasIdentifier: mediationExtrasIdentifier,
           extras: extras,
         );
@@ -634,6 +591,7 @@ class AdWidget extends StatefulWidget {
 class _AdWidgetState extends State<AdWidget> {
   bool _adIdAlreadyMounted = false;
   bool _adLoadNotCalled = false;
+  bool _firstVisibleOccurred = false;
 
   @override
   void initState() {
@@ -679,28 +637,51 @@ class _AdWidgetState extends State<AdWidget> {
       ]);
     }
     if (defaultTargetPlatform == TargetPlatform.android) {
-      return PlatformViewLink(
-        viewType: '${instanceManager.channel.name}/ad_widget',
-        surfaceFactory:
-            (BuildContext context, PlatformViewController controller) {
-          return AndroidViewSurface(
-            controller: controller as AndroidViewController,
-            gestureRecognizers: const <Factory<OneSequenceGestureRecognizer>>{},
-            hitTestBehavior: PlatformViewHitTestBehavior.opaque,
-          );
-        },
-        onCreatePlatformView: (PlatformViewCreationParams params) {
-          return PlatformViewsService.initSurfaceAndroidView(
-            id: params.id,
-            viewType: '${instanceManager.channel.name}/ad_widget',
-            layoutDirection: TextDirection.ltr,
-            creationParams: instanceManager.adIdFor(widget.ad),
-            creationParamsCodec: StandardMessageCodec(),
-          )
-            ..addOnPlatformViewCreatedListener(params.onPlatformViewCreated)
-            ..create();
-        },
-      );
+      // Do not attach the platform view widget until it will actually become
+      // visible. This is a workaround for
+      // https://github.com/googleads/googleads-mobile-flutter/issues/580,
+      // where impressions are erroneously fired due to how platform views are
+      // rendered.
+      // TODO (jjliu15): Remove this after the flutter issue is resolved.
+      if (_firstVisibleOccurred) {
+        return PlatformViewLink(
+          viewType: '${instanceManager.channel.name}/ad_widget',
+          surfaceFactory:
+              (BuildContext context, PlatformViewController controller) {
+            return AndroidViewSurface(
+              controller: controller as AndroidViewController,
+              gestureRecognizers: const <
+                  Factory<OneSequenceGestureRecognizer>>{},
+              hitTestBehavior: PlatformViewHitTestBehavior.opaque,
+            );
+          },
+          onCreatePlatformView: (PlatformViewCreationParams params) {
+            return PlatformViewsService.initSurfaceAndroidView(
+              id: params.id,
+              viewType: '${instanceManager.channel.name}/ad_widget',
+              layoutDirection: TextDirection.ltr,
+              creationParams: instanceManager.adIdFor(widget.ad),
+              creationParamsCodec: StandardMessageCodec(),
+            )
+              ..addOnPlatformViewCreatedListener(params.onPlatformViewCreated)
+              ..create();
+          },
+        );
+      } else {
+        final adId = instanceManager.adIdFor(widget.ad);
+        return VisibilityDetector(
+          key: Key('android-platform-view-$adId'),
+          onVisibilityChanged: (visibilityInfo) {
+            if (!_firstVisibleOccurred &&
+                visibilityInfo.visibleFraction > 0.01) {
+              setState(() {
+                _firstVisibleOccurred = true;
+              });
+            }
+          },
+          child: Container(),
+        );
+      }
     }
 
     return UiKitView(
@@ -856,15 +837,6 @@ class BannerAd extends AdWithView {
   /// A listener for receiving events in the ad lifecycle.
   @override
   final BannerAdListener listener;
-
-  /// Check out developer pages for [Admob](https://developers.google.com/admob/flutter/test-ads)
-  /// and [AdManager](https://developers.google.com/ad-manager/mobile-ads-sdk/flutter/test-ads) for
-  /// demo ad units that point to specific test creatives for each format.
-  @Deprecated(
-      'Use test ad unit ids from the developer page while creating the ad.')
-  static final String testAdUnitId = Platform.isAndroid
-      ? 'ca-app-pub-3940256099942544/6300978111'
-      : 'ca-app-pub-3940256099942544/2934735716';
 
   @override
   Future<void> load() async {
@@ -1034,15 +1006,6 @@ class NativeAd extends AdWithView {
   /// Options to configure the native ad request.
   final NativeAdOptions? nativeAdOptions;
 
-  /// Check out developer pages for [Admob](https://developers.google.com/admob/flutter/test-ads)
-  /// and [AdManager](https://developers.google.com/ad-manager/mobile-ads-sdk/flutter/test-ads) for
-  /// demo ad units that point to specific test creatives for each format.
-  @Deprecated(
-      'Use test ad unit ids from the developer page while creating the ad.')
-  static final String testAdUnitId = Platform.isAndroid
-      ? 'ca-app-pub-3940256099942544/2247696110'
-      : 'ca-app-pub-3940256099942544/3986624511';
-
   @override
   Future<void> load() async {
     await instanceManager.loadNativeAd(this);
@@ -1069,15 +1032,6 @@ class InterstitialAd extends AdWithoutView {
 
   /// Callbacks to be invoked when ads show and dismiss full screen content.
   FullScreenContentCallback<InterstitialAd>? fullScreenContentCallback;
-
-  /// Check out developer pages for [Admob](https://developers.google.com/admob/flutter/test-ads)
-  /// and [AdManager](https://developers.google.com/ad-manager/mobile-ads-sdk/flutter/test-ads) for
-  /// demo ad units that point to specific test creatives for each format.
-  @Deprecated(
-      'Use test ad unit ids from the developer page while creating the ad.')
-  static final String testAdUnitId = Platform.isAndroid
-      ? 'ca-app-pub-3940256099942544/1033173712'
-      : 'ca-app-pub-3940256099942544/4411468910';
 
   /// Loads an [InterstitialAd] with the given [adUnitId] and [request].
   static Future<void> load({
@@ -1158,7 +1112,6 @@ class RewardedAd extends AdWithoutView {
     required String adUnitId,
     required this.rewardedAdLoadCallback,
     required this.request,
-    this.serverSideVerificationOptions,
   })  : adManagerRequest = null,
         super(adUnitId: adUnitId);
 
@@ -1169,7 +1122,6 @@ class RewardedAd extends AdWithoutView {
     required String adUnitId,
     required this.rewardedAdLoadCallback,
     required this.adManagerRequest,
-    this.serverSideVerificationOptions,
   })  : request = null,
         super(adUnitId: adUnitId);
 
@@ -1182,18 +1134,6 @@ class RewardedAd extends AdWithoutView {
   /// Callbacks for events that occur when attempting to load an ad.
   final RewardedAdLoadCallback rewardedAdLoadCallback;
 
-  /// Check out developer pages for [Admob](https://developers.google.com/admob/flutter/test-ads)
-  /// and [AdManager](https://developers.google.com/ad-manager/mobile-ads-sdk/flutter/test-ads) for
-  /// demo ad units that point to specific test creatives for each format.
-  @Deprecated(
-      'Use test ad unit ids from the developer page while creating the ad.')
-  static final String testAdUnitId = Platform.isAndroid
-      ? 'ca-app-pub-3940256099942544/5224354917'
-      : 'ca-app-pub-3940256099942544/1712485313';
-
-  /// Optional [ServerSideVerificationOptions].
-  ServerSideVerificationOptions? serverSideVerificationOptions;
-
   /// Callbacks to be invoked when ads show and dismiss full screen content.
   FullScreenContentCallback<RewardedAd>? fullScreenContentCallback;
 
@@ -1205,13 +1145,11 @@ class RewardedAd extends AdWithoutView {
     required String adUnitId,
     required AdRequest request,
     required RewardedAdLoadCallback rewardedAdLoadCallback,
-    ServerSideVerificationOptions? serverSideVerificationOptions,
   }) async {
     RewardedAd rewardedAd = RewardedAd._(
         adUnitId: adUnitId,
         request: request,
-        rewardedAdLoadCallback: rewardedAdLoadCallback,
-        serverSideVerificationOptions: serverSideVerificationOptions);
+        rewardedAdLoadCallback: rewardedAdLoadCallback);
 
     await instanceManager.loadRewardedAd(rewardedAd);
   }
@@ -1221,13 +1159,11 @@ class RewardedAd extends AdWithoutView {
     required String adUnitId,
     required AdManagerAdRequest adManagerRequest,
     required RewardedAdLoadCallback rewardedAdLoadCallback,
-    ServerSideVerificationOptions? serverSideVerificationOptions,
   }) async {
     RewardedAd rewardedAd = RewardedAd._fromAdManagerRequest(
         adUnitId: adUnitId,
         adManagerRequest: adManagerRequest,
-        rewardedAdLoadCallback: rewardedAdLoadCallback,
-        serverSideVerificationOptions: serverSideVerificationOptions);
+        rewardedAdLoadCallback: rewardedAdLoadCallback);
 
     await instanceManager.loadRewardedAd(rewardedAd);
   }
@@ -1240,6 +1176,11 @@ class RewardedAd extends AdWithoutView {
   Future<void> show({required OnUserEarnedRewardCallback onUserEarnedReward}) {
     onUserEarnedRewardCallback = onUserEarnedReward;
     return instanceManager.showAdWithoutView(this);
+  }
+
+  /// Set [ServerSideVerificationOptions] for the ad.
+  Future<void> setServerSideOptions(ServerSideVerificationOptions options) {
+    return instanceManager.setServerSideVerificationOptions(options, this);
   }
 }
 
@@ -1260,7 +1201,6 @@ class RewardedInterstitialAd extends AdWithoutView {
     required String adUnitId,
     required this.rewardedInterstitialAdLoadCallback,
     required this.request,
-    this.serverSideVerificationOptions,
   })  : adManagerRequest = null,
         super(adUnitId: adUnitId);
 
@@ -1271,7 +1211,6 @@ class RewardedInterstitialAd extends AdWithoutView {
     required String adUnitId,
     required this.rewardedInterstitialAdLoadCallback,
     required this.adManagerRequest,
-    this.serverSideVerificationOptions,
   })  : request = null,
         super(adUnitId: adUnitId);
 
@@ -1283,9 +1222,6 @@ class RewardedInterstitialAd extends AdWithoutView {
 
   /// Callbacks for events that occur when attempting to load an ad.
   final RewardedInterstitialAdLoadCallback rewardedInterstitialAdLoadCallback;
-
-  /// Optional [ServerSideVerificationOptions].
-  ServerSideVerificationOptions? serverSideVerificationOptions;
 
   /// Callbacks to be invoked when ads show and dismiss full screen content.
   FullScreenContentCallback<RewardedInterstitialAd>? fullScreenContentCallback;
@@ -1299,13 +1235,11 @@ class RewardedInterstitialAd extends AdWithoutView {
     required AdRequest request,
     required RewardedInterstitialAdLoadCallback
         rewardedInterstitialAdLoadCallback,
-    ServerSideVerificationOptions? serverSideVerificationOptions,
   }) async {
     RewardedInterstitialAd rewardedInterstitialAd = RewardedInterstitialAd._(
         adUnitId: adUnitId,
         request: request,
-        rewardedInterstitialAdLoadCallback: rewardedInterstitialAdLoadCallback,
-        serverSideVerificationOptions: serverSideVerificationOptions);
+        rewardedInterstitialAdLoadCallback: rewardedInterstitialAdLoadCallback);
 
     await instanceManager.loadRewardedInterstitialAd(rewardedInterstitialAd);
   }
@@ -1316,15 +1250,13 @@ class RewardedInterstitialAd extends AdWithoutView {
     required AdManagerAdRequest adManagerRequest,
     required RewardedInterstitialAdLoadCallback
         rewardedInterstitialAdLoadCallback,
-    ServerSideVerificationOptions? serverSideVerificationOptions,
   }) async {
     RewardedInterstitialAd rewardedInterstitialAd =
         RewardedInterstitialAd._fromAdManagerRequest(
             adUnitId: adUnitId,
             adManagerRequest: adManagerRequest,
             rewardedInterstitialAdLoadCallback:
-                rewardedInterstitialAdLoadCallback,
-            serverSideVerificationOptions: serverSideVerificationOptions);
+                rewardedInterstitialAdLoadCallback);
 
     await instanceManager.loadRewardedInterstitialAd(rewardedInterstitialAd);
   }
@@ -1337,6 +1269,11 @@ class RewardedInterstitialAd extends AdWithoutView {
   Future<void> show({required OnUserEarnedRewardCallback onUserEarnedReward}) {
     onUserEarnedRewardCallback = onUserEarnedReward;
     return instanceManager.showAdWithoutView(this);
+  }
+
+  /// Set [ServerSideVerificationOptions] for the ad.
+  Future<void> setServerSideOptions(ServerSideVerificationOptions options) {
+    return instanceManager.setServerSideVerificationOptions(options, this);
   }
 }