+++ packages/google_mobile_ads/test/ad_containers_test.dart
@@ -18,9 +18,9 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:google_mobile_ads/src/ad_instance_manager.dart';
 import 'package:google_mobile_ads/google_mobile_ads.dart';
-import 'package:flutter/widgets.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:flutter/services.dart';
+import 'package:visibility_detector/visibility_detector.dart';
 
 // ignore_for_file: deprecated_member_use_from_same_package
 void main() {
@@ -49,6 +49,7 @@ void main() {
           case 'loadInterstitialAd':
           case 'loadAdManagerInterstitialAd':
           case 'loadAdManagerBannerAd':
+          case 'setServerSideVerificationOptions':
             return Future<void>.value();
           case 'getAdSize':
             return Future<dynamic>.value(AdSize.banner);
@@ -102,21 +103,18 @@ void main() {
       ]);
     });
 
-    test('load rewarded ad and set immersive mode', () async {
+    test('load rewarded ad and set immersive mode and ssv', () async {
       RewardedAd? rewarded;
       AdRequest request = AdRequest();
       await RewardedAd.load(
-          adUnitId: 'test-ad-unit',
-          request: request,
-          rewardedAdLoadCallback: RewardedAdLoadCallback(
-              onAdLoaded: (ad) {
-                rewarded = ad;
-              },
-              onAdFailedToLoad: (error) => null),
-          serverSideVerificationOptions: ServerSideVerificationOptions(
-            userId: 'test-user-id',
-            customData: 'test-custom-data',
-          ));
+        adUnitId: 'test-ad-unit',
+        request: request,
+        rewardedAdLoadCallback: RewardedAdLoadCallback(
+            onAdLoaded: (ad) {
+              rewarded = ad;
+            },
+            onAdFailedToLoad: (error) => null),
+      );
 
       RewardedAd createdAd = instanceManager.adFor(0) as RewardedAd;
       (createdAd).rewardedAdLoadCallback.onAdLoaded(createdAd);
@@ -127,19 +125,27 @@ void main() {
           'adUnitId': 'test-ad-unit',
           'request': request,
           'adManagerRequest': null,
-          'serverSideVerificationOptions':
-              rewarded!.serverSideVerificationOptions,
         }),
       ]);
 
       expect(instanceManager.adFor(0), isNotNull);
       expect(rewarded, createdAd);
 
+      // Set immersive mode
       log.clear();
       await createdAd.setImmersiveMode(true);
       expect(log, <Matcher>[
         isMethodCall('setImmersiveMode',
             arguments: {'adId': 0, 'immersiveModeEnabled': true})
+      ]);
+
+      // Set ssv
+      log.clear();
+      final ssv = ServerSideVerificationOptions();
+      await createdAd.setServerSideOptions(ssv);
+      expect(log, <Matcher>[
+        isMethodCall('setServerSideVerificationOptions',
+            arguments: {'adId': 0, 'serverSideVerificationOptions': ssv})
       ]);
     });
 
@@ -276,7 +282,8 @@ void main() {
       expect(instanceManager.adFor(0), isNotNull);
     });
 
-    testWidgets('build ad widget', (WidgetTester tester) async {
+    testWidgets('build ad widget iOS', (WidgetTester tester) async {
+      debugDefaultTargetPlatformOverride = TargetPlatform.iOS;
       final NativeAd native = NativeAd(
         adUnitId: 'test-ad-unit',
         factoryId: '0',
@@ -287,41 +294,89 @@ void main() {
       await native.load();
 
       await tester.pumpWidget(
-        Builder(
-          builder: (BuildContext context) {
-            AdWidget widget = AdWidget(ad: native);
-            Widget buildWidget = widget.createElement().build();
-            expect(buildWidget, isA<PlatformViewLink>());
-            return widget;
-          },
+        MaterialApp(
+          home: Material(
+            child: SingleChildScrollView(
+              child: Column(
+                key: UniqueKey(),
+                children: [
+                  SizedBox.fromSize(size: Size(200, 1000)),
+                  Container(
+                    height: 200,
+                    width: 200,
+                    child: AdWidget(ad: native),
+                  ),
+                ],
+              ),
+            ),
+          ),
         ),
       );
+      await tester.pumpAndSettle();
+
+      final uiKitView = tester.widget(find.byType(UiKitView));
+      expect(uiKitView, isNotNull);
 
       await native.dispose();
+      debugDefaultTargetPlatformOverride = null;
     });
 
-    testWidgets('build ad widget', (WidgetTester tester) async {
-      final NativeAd native = NativeAd(
+    testWidgets('Build ad widget Android', (WidgetTester tester) async {
+      debugDefaultTargetPlatformOverride = TargetPlatform.android;
+      // Create a loaded ad
+      final ad = NativeAd(
         adUnitId: 'test-ad-unit',
         factoryId: '0',
         listener: NativeAdListener(),
         request: AdRequest(),
       );
+      await ad.load();
 
-      await native.load();
-
+      // Render ad in a scrolling view
+      VisibilityDetectorController.instance.updateInterval = Duration.zero;
       await tester.pumpWidget(
-        Builder(
-          builder: (BuildContext context) {
-            AdWidget widget = AdWidget(ad: native);
-            Widget buildWidget = widget.createElement().build();
-            expect(buildWidget, isA<PlatformViewLink>());
-            return widget;
-          },
+        MaterialApp(
+          home: Material(
+            child: SingleChildScrollView(
+              child: Column(
+                key: UniqueKey(),
+                children: [
+                  SizedBox.fromSize(size: Size(200, 1000)),
+                  Container(
+                    height: 200,
+                    width: 200,
+                    child: AdWidget(ad: ad),
+                  ),
+                ],
+              ),
+            ),
+          ),
         ),
       );
+      await tester.pumpAndSettle();
 
-      await native.dispose();
+      // On initial render, VisibilityRender should be in the UI
+      final visibilityDetectorWidget =
+          tester.widget(find.byKey(Key('android-platform-view-0')));
+      expect(visibilityDetectorWidget, isNotNull);
+      expect(visibilityDetectorWidget, isA<VisibilityDetector>());
+      final platformViewLinks =
+          tester.widgetList(find.byType(PlatformViewLink));
+      expect(platformViewLinks.isEmpty, true);
+
+      // Drag the ad widget into view
+      await tester.drag(find.byType(SingleChildScrollView), Offset(0.0, -1000));
+      await tester.pumpAndSettle();
+
+      // PlatformViewLink should now be present instead of VisibilityDetector
+      final detectors = tester.widgetList(find.byType(VisibilityDetector));
+      expect(detectors.isEmpty, true);
+      final platformViewLink = tester.widget(find.byType(PlatformViewLink));
+      expect(platformViewLink, isNotNull);
+
+      // Reset platform override
+      await ad.dispose();
+      debugDefaultTargetPlatformOverride = null;
     });
 
     testWidgets('warns when ad has not been loaded',
@@ -488,17 +543,14 @@ void main() {
       RewardedAd? rewarded;
       AdRequest request = AdRequest();
       await RewardedAd.load(
-          adUnitId: 'test-ad-unit',
-          request: request,
-          rewardedAdLoadCallback: RewardedAdLoadCallback(
-              onAdLoaded: (ad) {
-                rewarded = ad;
-              },
-              onAdFailedToLoad: (error) => null),
-          serverSideVerificationOptions: ServerSideVerificationOptions(
-            userId: 'test-user-id',
-            customData: 'test-custom-data',
-          ));
+        adUnitId: 'test-ad-unit',
+        request: request,
+        rewardedAdLoadCallback: RewardedAdLoadCallback(
+            onAdLoaded: (ad) {
+              rewarded = ad;
+            },
+            onAdFailedToLoad: (error) => null),
+      );
 
       RewardedAd createdAd = instanceManager.adFor(0) as RewardedAd;
       (createdAd).rewardedAdLoadCallback.onAdLoaded(createdAd);
@@ -509,8 +561,6 @@ void main() {
           'adUnitId': 'test-ad-unit',
           'request': request,
           'adManagerRequest': null,
-          'serverSideVerificationOptions':
-              rewarded!.serverSideVerificationOptions,
         }),
       ]);
 
@@ -530,17 +580,14 @@ void main() {
       RewardedAd? rewarded;
       AdManagerAdRequest request = AdManagerAdRequest();
       await RewardedAd.loadWithAdManagerAdRequest(
-          adUnitId: 'test-ad-unit',
-          adManagerRequest: request,
-          rewardedAdLoadCallback: RewardedAdLoadCallback(
-              onAdLoaded: (ad) {
-                rewarded = ad;
-              },
-              onAdFailedToLoad: (error) => null),
-          serverSideVerificationOptions: ServerSideVerificationOptions(
-            userId: 'test-user-id',
-            customData: 'test-custom-data',
-          ));
+        adUnitId: 'test-ad-unit',
+        adManagerRequest: request,
+        rewardedAdLoadCallback: RewardedAdLoadCallback(
+            onAdLoaded: (ad) {
+              rewarded = ad;
+            },
+            onAdFailedToLoad: (error) => null),
+      );
 
       RewardedAd createdAd = instanceManager.adFor(0) as RewardedAd;
       (createdAd).rewardedAdLoadCallback.onAdLoaded(createdAd);
@@ -551,8 +598,6 @@ void main() {
           'adUnitId': 'test-ad-unit',
           'request': null,
           'adManagerRequest': request,
-          'serverSideVerificationOptions':
-              rewarded!.serverSideVerificationOptions,
         }),
       ]);
 
@@ -803,7 +848,6 @@ void main() {
           'adUnitId': 'test-ad-unit',
           'request': request,
           'adManagerRequest': null,
-          'serverSideVerificationOptions': null,
         })
       ]);
 
@@ -902,11 +946,7 @@ void main() {
               onAdLoaded: (ad) {
                 rewarded = ad;
               },
-              onAdFailedToLoad: (error) => null),
-          serverSideVerificationOptions: ServerSideVerificationOptions(
-            userId: 'test-user-id',
-            customData: 'test-custom-data',
-          ));
+              onAdFailedToLoad: (error) => null));
 
       RewardedAd createdAd = instanceManager.adFor(0) as RewardedAd;
       createdAd.rewardedAdLoadCallback.onAdLoaded(createdAd);
@@ -1065,8 +1105,6 @@ void main() {
         nonPersonalizedAds: false,
         neighboringContentUrls: <String>['url1.com', 'url2.com'],
         httpTimeoutMillis: 12345,
-        location:
-            LocationParams(accuracy: 1.1, longitude: 25, latitude: 38, time: 1),
         mediationExtrasIdentifier: 'identifier',
         extras: {'key': 'value'},
       );
@@ -1093,8 +1131,6 @@ void main() {
         nonPersonalizedAds: false,
         neighboringContentUrls: <String>['url1.com', 'url2.com'],
         httpTimeoutMillis: 12345,
-        location:
-            LocationParams(accuracy: 1.1, longitude: 25, latitude: 38, time: 1),
         mediationExtrasIdentifier: 'identifier',
         extras: {'key': 'value'},
       );
@@ -1106,7 +1142,6 @@ void main() {
       expect(decoded.contentUrl, adRequest.contentUrl);
       expect(decoded.nonPersonalizedAds, adRequest.nonPersonalizedAds);
       expect(decoded.keywords, adRequest.keywords);
-      expect(decoded.location, null);
       expect(decoded.mediationExtrasIdentifier, 'identifier');
     });
 
@@ -1257,8 +1292,6 @@ void main() {
         neighboringContentUrls: <String>['url1.com', 'url2.com'],
         httpTimeoutMillis: 5000,
         publisherProvidedId: 'test-pub-id',
-        location:
-            LocationParams(accuracy: 1.1, longitude: 25, latitude: 38, time: 1),
         mediationExtrasIdentifier: 'identifier',
         extras: {'key': 'value'},
       );
@@ -1298,8 +1331,6 @@ void main() {
         neighboringContentUrls: <String>['url1.com', 'url2.com'],
         httpTimeoutMillis: 5000,
         publisherProvidedId: 'test-pub-id',
-        location:
-            LocationParams(accuracy: 1.1, longitude: 25, latitude: 38, time: 1),
         mediationExtrasIdentifier: 'identifier',
         extras: {'key': 'value'},
       );
@@ -1314,20 +1345,17 @@ void main() {
       expect(decoded.publisherProvidedId, request.publisherProvidedId);
       expect(decoded.customTargeting, request.customTargeting);
       expect(decoded.customTargetingLists, request.customTargetingLists);
-      expect(decoded.location, null);
       expect(decoded.mediationExtrasIdentifier, 'identifier');
     });
 
     test('ad click native', () async {
       var testNativeClick = (eventName, adId) async {
-        final Completer<Ad> nativeAdClickCompleter = Completer<Ad>();
         final Completer<Ad> adClickCompleter = Completer<Ad>();
 
         final NativeAd native = NativeAd(
           adUnitId: 'test-ad-unit',
           factoryId: 'testId',
           listener: NativeAdListener(
-              onNativeAdClicked: (Ad ad) => nativeAdClickCompleter.complete(ad),
               onAdClicked: (ad) => adClickCompleter.complete(ad)),
           request: AdRequest(),
         );
@@ -1346,7 +1374,6 @@ void main() {
           (ByteData? data) {},
         );
 
-        expect(nativeAdClickCompleter.future, completion(native));
         expect(adClickCompleter.future, completion(native));
       };
 
@@ -1373,11 +1400,7 @@ void main() {
                   ad.fullScreenContentCallback = FullScreenContentCallback(
                       onAdClicked: (ad) => adClickCompleter.complete(ad));
                 },
-                onAdFailedToLoad: (error) => null),
-            serverSideVerificationOptions: ServerSideVerificationOptions(
-              userId: 'test-user-id',
-              customData: 'test-custom-data',
-            ));
+                onAdFailedToLoad: (error) => null));
 
         MethodCall methodCall = MethodCall('onAdEvent', <dynamic, dynamic>{
           'adId': adId,